+++ libs/pavex/src/blueprint/middleware/mod.rs
@@ -4,8 +4,9 @@
 //!
 //! Check out the ["Middleware"](https://pavex.dev/docs/guide/middleware) section of Pavex's guide
 //! for a thorough introduction to middlewares in Pavex applications.
-mod registered;
-mod unregistered;
+mod post;
+mod pre;
+mod wrapping;
 
-pub use registered::{RegisteredPostProcessingMiddleware, RegisteredWrappingMiddleware};
-pub use unregistered::{PostProcessingMiddleware, WrappingMiddleware};
+pub use post::{PostProcessingMiddleware, RegisteredPostProcessingMiddleware};
+pub use wrapping::{RegisteredWrappingMiddleware, WrappingMiddleware};
