+++ libs/pavex/src/blueprint/middleware/post/mod.rs
@@ -1,0 +1,5 @@
+mod registered;
+mod unregistered;
+
+pub use registered::RegisteredPostProcessingMiddleware;
+pub use unregistered::PostProcessingMiddleware;
