+++ libs/pavex/src/blueprint/middleware/wrapping/mod.rs
@@ -1,0 +1,5 @@
+mod registered;
+mod unregistered;
+
+pub use registered::RegisteredWrappingMiddleware;
+pub use unregistered::WrappingMiddleware;
